% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cor.R
\name{color_cor}
\alias{color_cor}
\title{Background colored correlation plot of two columns of data}
\usage{
color_cor(data, mapping, color = I("black"), size = 6, palette = "RdYlGn",
  breaks = c(-0.99, -0.8, -0.6, 0.6, 0.8, 0.99), ..., use = "everything",
  method = "pearson", reversePalette = FALSE)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used. Must contain 'x' and 'y'. 'color' will be ignored}

\item{color}{color of main text. Default = "black"}

\item{size}{ggplot2 size of main text. Default = 6}

\item{palette}{should be a character string that matches a RColorBrewer color palette}

\item{breaks}{can be a vector of breaks points between -1 and 1.  If a single numeric number, will set breaks with that many groups}

\item{...}{other arguments being supplied to ggally_text}

\item{use}{\code{use} supplied to cor function}

\item{method}{\code{method} suppied to cor function}

\item{reversePalette}{boolean to determine if the palette should be reversed}
}
\description{
Display the correlation of the given data and
}
\examples{
color_cor(iris, ggplot2::aes(x = Sepal.Length, y = Petal.Width))
color_cor(
  iris,
  ggplot2::aes(x = Sepal.Length, y = -1 * Petal.Width),
  method = "spearman"
)

require(GGally)
ggpairs(iris, 1:4, upper = list(continuous = color_cor))
# use kendall correlation, only three color groups and reverse the palette
ggpairs(
  iris, 1:3,
  lower = "blank",
  upper = list(
    continuous = wrap(
      color_cor,
      method = "kendall",
      breaks = 3,
      reversePalette = TRUE
    )
  )
)
# use spearman correlation, only three color groups but breaking at abs(0.2)
ggpairs(
  iris, 1:3,
  lower = "blank",
  upper = list(
    continuous = wrap(
      color_cor,
      method = "spearman",
      breaks = c(-0.20, 0.20)
    )
  )
)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

