% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cor.R
\name{ggally_cor_color}
\alias{ggally_cor_color}
\title{Background colored correlation plot of two columns of data}
\usage{
ggally_cor_color(data, mapping, color = I("black"), size = 6,
  pallete = "RdYlGn", breaks = c(-0.99, -0.8, -0.6, 0.6, 0.8, 0.99), ...,
  use = "everything", method = "pearson", reversePalette = FALSE)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used. Must contain 'x' and 'y'. 'color' will be ignored}

\item{color}{color of main text. Default = "black"}

\item{size}{ggplot2 size of main text. Default = 6}

\item{pallete}{should be a character string that matches a RColorBrewer color pallete}

\item{breaks}{can be a vector of breaks points between -1 and 1.  If a single numeric number, will set breaks with that many groups}

\item{...}{other arguments being supplied to ggally_text}

\item{use}{\code{use} supplied to cor function}

\item{method}{\code{method} suppied to cor function}

\item{reversePalette}{boolean to determine if the palette should be reversed}
}
\description{
Display the correlation of the given data and
}
\examples{
ggally_cor_color(iris, ggplot2::aes(x = Sepal.Length, y = Petal.Width))
ggally_cor_color(
  iris,
  ggplot2::aes(x = Sepal.Length, y = -1 * Petal.Width),
  method = "spearman"
)

require(GGally)
ggpairs(iris, 1:4, upper = list(continuous = ggally_cor_color))
# use kendall correlation, only three color groups and reverse the pallete
ggpairs(
  iris, 1:4,
  upper = list(
    continuous = wrap(
      ggally_cor_color,
      method = "kendall",
      breaks = 3,
      reversePalette = TRUE
    )
  )
)
# use kendall correlation, only three color groups but breaking at abs(0.2)
ggpairs(
  iris, 1:4,
  upper = list(
    continuous = wrap(
      ggally_cor_color,
      method = "spearman",
      breaks = c(-0.20, 0.20)
    )
  )
)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

